+++ src/services/nativeAuth/methods/index.ts
@@ -1,3 +1,2 @@
 export * from './getTokenExpiration';
 export * from './getNativeAuthConfig';
-export * from './getToken';